Moves project map into it's own tab. re #111
