+++ src/main.js
@@ -17,6 +17,7 @@ import underscore from 'underscore';
 import PageHeader from './app/shared/components/PageHeader.vue';
 import PageHeaderLayout from './app/shared/components/PageHeaderLayout.vue';
 import ProjectSummaryPage from './app/projects/components/ProjectSummaryPage.vue';
+import ProjectMapPage from './app/projects/components/ProjectMapPage.vue';
 import SelectResourceInstancePage from './app/projects/components/SelectResourceInstancePage.vue';
 import SelectResourceTypePage from './app/projects/components/SelectResourceTypePage.vue';
 import ResourceEditPage from './app/editor/components/ResourceEditPage.vue';
@@ -25,6 +26,7 @@ import StringWidget from './app/widgets/StringWidget.vue';
 Vue.component('page-header', PageHeader);
 Vue.component('page-header-layout', PageHeaderLayout);
 Vue.component('project-summary-page', ProjectSummaryPage);
+Vue.component('project-map-page', ProjectMapPage);
 Vue.component('select-resource-instance-page', SelectResourceInstancePage);
 Vue.component('select-resource-type-page', SelectResourceTypePage);
 Vue.component('resource-edit-page', ResourceEditPage);
